Check in a skeleton based on the ionic starter project
This was generated using the yeoman ionic generator.
https://github.com/diegonetto/generator-ionic

so unlike the classic ionic/cordova, all the real code is under "app" and is
minifed before being copied into www

The only change to the starter project was to change the "img" to "images" in
the service to make the images load correctly.