Add a login screen that logs in via the native code implementations
This is actually fairly straightforward. I am trying to replicate the toast
that describes the login like we used to have on android, but that doesn't
really seem to work and I am not sure why.

The timeout is to allow enough time for the toast to be seen. If we don't end
up going with the toast, we can just skip timeout and move right along.