Finally get a fully nested view tree to work
There is now a root abstract node that resolves the cordova plugins.
Everything else depends on it.

This is good because then we know that there is a place to put javascript init
code. If we move the initialization out to individual plugins and don't need a
single place to init in here, we can revisit this decision.

Once the intro is done, we store a shared preference stating that it is done.

And in the root scope, we automatically choose the intro or the main view
depending on whether we have previously finished the intro or not.