Get the screens to work on android as well
The documentation for the ui-router states that an abstract parent that is used
primarily for a resolve needs to have a basic <ui-view/> template. However,
this assumes the use of standard angular.

In ionic, all the angular CSS and directives are not defined, so ui-view is not defined.
This works on ios, but fails on android.

Note that replacing it with ion-view *does not work*. In ionic, the element
that is augmented with the routes is `ion-nav-view` and `ion-nav-view` alone.