Turns on the main diary screen
All the dependencies are in place and everything seems to work.
I haven't actually tested it against any real data, since the existing database
is nuked when we init the usercache.

Also, it looks like we show one list item even when there are no trips, and we
skip the summary at the top. Need to improve testing, but let's check in what
we've got, which is almost all of it.