Add a button to the control screen to return to the intro and sign in again
Without this, if we login using the developer login, then when the application
was restarted, then the login info would be deleted, since it was not stored as
a cookie in the keychain. We should really store the username as a preference
in the native code, but this is a good workaround for now.