Retrieve data from the server if it doesn't exist in the usercache
We no longer retrieve the data from the local development server started with
"ionic serve". Instead, we retrieve the data from the real server. This is
important if we want to treat the usercache as a true cache instead of just
storing all the data in there.

So we use the BEMServerComm plugin to get the appropriate user specific data.
Note that since this data is returned from bottle, it needs to be a dict and
not an array, so it is a dict with one key "timeline" and the array as the
value.

With this in place, we can now see a reasonable timeline.

We just need to add the fixes to leaflet and the angular nvd3 directive, and
fix the CORS stuff on iOS.