Add more logging to help debug WSODs
In spite of all my fixes, we still occasionally get a WSOD on iOS.
And in spite of all the attempts to be able to see the result in Safari, I am
not able to do so.

Let's add some more debugging messages to help dignose the issue when it does
occur.