Remove the old and obsolete "app" directory
This is a legacy from the time we tried to use the ionic yeoman generator as a
starter project. Unfortunately, that project was old and tried to install old
versions of cordova. It was also not consistent with the current ionic app
structures.

So I moved everything out to www, but left the "app" directory around just in case.
Now it's time to remove it too.