Fix scrolling on top of the map in android
I've documented this in the HTML file as well, but we need to repeat using
collection-repeat, not ng-repeat.

Otherwise, you can only scroll using the sidebar, scrolling on the map
side does not work.  I have no clue why this happens, but it is not worth it to
understand this deeply. It is not purely a leaflet issue, but it also clearly
something to do with leaflet because scrolling on the details part does work.