Finally fix WSOD for once and for all
I think that part of the reason that the WSOD issue was so hard to fix is that
there were multiple causes.

- race between splashscreen plugin and the device plugin causes device ready to
  never fire (https://issues.apache.org/jira/browse/CB-10530), fixed in
cordova-ios@4.1.0.
- issues with races between config and run if we waited for device ready and
  then started loading
- if the initial load failed, we never retry

Fixed all of these by adding a new splash screen and splash state that is
loaded by default. The splash screen then loads the user preferred state in a
loop until the load completes.

This seems to work pretty reliably, after moving the pluginInit of the data
collector to its own thread, I don't get a WSOD even on the initial load, which
used to fail pretty reliably.

And because I try until I find the plugin for the appPreferences, all subequent
code can use native plugins without any problems.

Because of this, I was able to restore the original principled loading.

I believe that this fixes both:
https://github.com/e-mission/e-mission-phone/issues/12
and
https://github.com/e-mission/e-mission-phone/issues/19

Yay!