Revert change to reload the page on background sync
It turns out that while this was a good idea, it doesn't work properly in the
background. In particular, on iOS, plugins are not loaded if backgroundFetch or
remote push are invoked, since they don't initialize the app. On android, it
looks like the thread ends before the maps are fully loaded, so we have half
displayed, frozen maps. We should really check the status, reload here if
active and reload everything normally on resume. For now, we just add a refresh
button to avoid maintaining status.