Attempt to add popups to the places and trips
Unfortunately, the popups don't work. I tried switching to a click event and
the click event did not work. I tried listening to generic click events,
similar to the example and they didn't work.
http://angular-ui.github.io/ui-leaflet/#!/examples/events

I tried looking at the leaflet geojson example from the website in the simulator. It worked.
I tried adding breakpoints in that code and in my code.

Basically, in the working code, `_onUp` calls `_simulateEvent`

```
    // simulate click if the touch didn't move too much
    if (this._isTapValid()) {
        this._simulateEvent('click', first);
    }
```

which calls `e.target.dispatchEvent`, which calls `L.FeatureGroup._propagateEvent`.

In the non-working code, the target is different from the working code -
instead it being a `path.leaflet-clickable`, it is an icon. And then the
_propagateEvent is never invoked.

Punting on that until the next release.