Get the master-detail view for the trip diary to finally work
Woot! This is awesome! The changes were actually fairly subtle - the moral of
the story is that you have to follow the standard pattern VERY CAREFULLY. Blind
copy-pasting Does Not Work.

In particular, the view for the detail screen needs to be the same as the view
for the diary because they are in the same tab. Otherwise, ionic tries to load
the tab for the detail view, which fails in the general case because there is
no tab for the detail view. This is also why the $state.go code did not work
until we added the tab.

I suspect that now, $state.go will work, and so will the original formulation
of the state. But I also made the following changes to make it consistent with
the chats while trying to get it to work, and I'm not sure it is worth it to
revert them.

I changed the state to diary-detail instead of diary.detail so that it is not a
child of diary. This means that the URL for the detail state can be largely
the same as the diary with a suffix for the trip ID - i.e. '/diary/:tripId'
instead of 'detail/:tripId'. I am not sure if it matters, but the code is not
ugly, so I'm going to let it stand since it looks like this is fairly fragile.