Gracefully handle the case in which common trips are not available
The previous checkin assumes that the common trips object is always available.
Since it is created on every run of the pipeline, and synced to the phone
automatically, this is almost always true.

But there can still be a subtle race condition in which you have installed the
app, but the first sync has not happened. In that case, the common trips object
will not exist, and we will get a parse error while parsing it.

We now handle this case by creating a dummy object in that case. As a bonus,
this will also handle the case in which the object exists, but is malformed.