Change the tilemap for the app maps from OSM to stamen toner
We have been using the OSM tilemap for now, which is the default and is easy
to use. Unfortunately, since OSM is a free, volunteer service, they are not
really able to support CPU and bandwidth requirements for a large number of
tiles.

http://wiki.openstreetmap.org/wiki/Tile_usage_policy

We could switch to one of the free with API key or paid tile providers, but
that will require some investigation on which one to pick. We switch to the
stamen toner, tiles instead, which are free to use and do not require an API
key.

While doing that, also refactor the code so that the tiles location is
configured in one place. This should make it easy to change later or maybe
even allow the user to change the map background that she wants - watercolor?
toner lite?