Removed 2 hours and 6 hours since we can't support them and remain backwards compatible
The current iOS phones are all registered to no channel and push notifications
are broadcast to all phones. They won't register to the default `interval_3600`
channel unless people update, and it is not clear that they will. So in order
to maintain backwards compatibility, we need to continue broadcasting push
notifications every hour until everybody migrates over.

This means that we can't really support 2 hour and 4 hour cadences, since all
phones will receive updates every hour anyway. So we remove those entries from
the list of valid options, for now.