Restructure the common trips screen to be more complex
The default view is still a map, but we also have the option of seeing a list
or details of the places.

Similar to the recent view, the selection between screens is using a sidebar
menu. I would have liked to use tabs, but ionic tabs cannot be used inside a
content, or at least, I could not get it to work.

http://ionicframework.com/docs/api/directive/ionTabs/

> Note: do not place ion-tabs inside of an ion-content element; it has been known to cause a certain CSS bug.

The actual functionality in list and detail have not yet been implemented, but
I wanted to get the structure in place first for easy reviewing.