Start adding trip information
Display the number of trips for each successor.
Note that this was an assert, which unfortunately, triggered.

```
+        var retVal = null;
+        var maxCount = 0;
+        retArray.forEach(function(cTrip, index, array) {
+          if (cTrip.trips.length > maxCount) {
+            maxCount = cTrip.trips.length;
+            retVal = cTrip;
+          }
+        });
```

On further introspection, it looks like there are indeed multiple common trips
with the same end points but different durations, I am not sure why this is
happening. But since we will probably change the common graph calculation
again, let's punt on this for now and take the common trip with the most
normal trips matched to it.