Added some additional parsing for the common trips information
Added some useful functions to find the highest frequency entries instead of
the average entries and refactored some existing code to use it.

Also renamed the findCommon() method to trip2Common() and added a similar
function for place. We can also add additional time_fns to get the most common
entry and exit times and so on.

Also changed all instances of the old getCommon to the new function to prevent
breakage.

```
bash-3.2$ grep -r findCommon www/
bash-3.2$
```