Add a new method to the `CommHelper` to invoke the habitica proxy API on our server
Again, we use the `CommHelper` so that we can add the auth headers
automatically and make an authenticated user call.
The server implementation is at
https://github.com/e-mission/e-mission-server/pull/307

sample `callObj` objects are in the comments before the function.
Note that this is completely untested.

In particular, I am not sure whether the JSON generated using `null` objects
for the args can be serialized correctly. If this does not work, we should skip
the args for `GET` messages, and change the server-side to check whether the
`method_args` key exists before reading it and pass in None if it doesn't.

However, I assume that @sunil07t and @juemura can make such minor fixes if the
basic structure is in place.