Add a button to manually call 'invalidateSize' on the map
As tracked in
https://github.com/e-mission/e-mission-phone/issues/75
on android, the map is frequently only partially visible.
Force killing the app fixes the issue, but until that happens, it looks ugly
and stupid.

The issue may be related to
https://github.com/Leaflet/Leaflet/issues/694
and fixed by calling `invalidateSize()` on the map, but it may also be some
other weirdness with using Leaflet on mobile.

So we add a button to `$broadcast 'invalidateSize'$ that can be used when the
weird behaviour occurs. If this fixes it, we can figure out how to call
`'invalidateSize'` automatically.