Restore non-crosswalk apk creation for L+
Back in June, we had tried to slim down the apk by:
(1) using crosswalk lite
(2) using the system webview for android 5.0+, which has a system webview

It turned out that on devices with android 5.0+, the control screen for the
heatmap was not displayed correctly.
https://github.com/e-mission/e-mission-phone/issues/78

We thought that this was because of differences between the system webview and
crosswalk - (2) above - so I reverted that change as part of commit
fb6362a90b24ba917b0bc9cd80f1ca6af9129d6f

But some testing by @yw374cornell established that the controls are not fully
visible even with crosswalk enabled. And with checking in all the images for
habitica, the apk size is now 50+MB! This will still work on google play, but
is not too great.

Let's restore the use of the system webview, at least for 5.0+ devices...