File tile loading on android
Apparently because of loading differences between android and iOS, the leaflet
maps on android frequently ended up with tiles only partially loaded.
This is tracked in issue:
https://github.com/e-mission/e-mission-phone/issues/75

In e3d6f7181c25a4ec0742f949854c7b9c77e5794a, I added a button to manually
apply the workaround for the issue and @yw374cornell confirmed that it worked.

In this change, I apply the workaround automatically and for all maps.

There are also minor whitespace differences that were automatically added by
the editor configuration. Anything that is not related to the leafletDirective.*resize code is whitespace.

Also added code to listen to all events in the common map so that we can get a
better sense of what events are actually generated on the phone versus in the
emulator. Since we do not have access to an android phone or emulator, it can
also help us debug in case android does not generate resize events and the
workaround still does not work.