Add an angular service wrapper over the unified logger
This was originally to log javascript messages both to the console and to the
unified logger, but I have now changed the unified logger to support this
natively. But it is still good to potentially cache log messages when native
plugins are not ready, so that we don't have to wait for native code during
startup.