Add support for re-consent
Requirements are tracked in
https://github.com/e-mission/e-mission-phone/issues/43

Current functionality:
- if intro is not complete, start intro
- if intro is complete, but there's a new consent version, reconsent
- if intro is complete, and current version is consented, move on

Further, if the new version has not yet been consented to, then the user cannot
access user specific tabs such as "diary" and "common. If she attemps to do so,
she is presented with the consent screen again. She can, however, access
generic tabs such as "heatmap" and "metrics".

Consent is written to:
- usercache (for use with native code)
- local storage (for use with subsequent javascript sessions)
- $rootScope (for use with current javascript session)

Further the splash controller now has a `'stateChangeStart'` listener which
intercepts state change events and displays the consent screen if necessary. It
also adds listeners for various error conditions to help debugging issues with
state definitions.