Clone queries before executing them
This is a fix for https://github.com/e-mission/e-mission-phone/issues/93
Basically, while issuing multiple calls to the server to retrieve result
metrics, @krisma reused the data object. Since javascript passes by reference
instead of by copy, this meant that the arguments object was modified in
place, and instead of four queries with different parameters, we ended up with four queries for the same parameter.

Fixed it by cloning the args before making the queries