Replicate the solution to "fix tiles" to all screens with maps
While trying to fix
https://github.com/e-mission/e-mission-phone/issues/75
the solution of listening to map resize events and then invalidating the map
did not work. However, the brute force solution of adding a button to the
screen did work. Let's replicate that solution multiple times to get this
release out of the door right now, and fix it the right way later.