Fix group create, invite and login
Fixes several minor issues around group create, invite and login to fix issues
found when SDB tried to log in. What happened was that people tried to create
eroups, but the party screen did not refresh. On manually refreshing the
screen, the parties showed up, but when they invited others, the invitations
didn't show up. And when they logged in via the UI, they got an error saying
that the group was invalid.

I was able to reproduce the error by using "Create party" without entering a
party name. In this case, the party name that we use for creation is
"undefined", and it looks like that causes the party creation in habitica to
get into some kind of wierd state where a party ID exists, but the party is not
actually valid or something.

I fixed this by checking whether a party name exists, and forcing you to enter
it if it does. In general, if there are errors in making remote calls, we
should display them instead of just logging them to the console where they are
dead and buried.

Also, if we attempted to fix this by deleting the user from habitica, the
screen wouldn't return to login even on refresh. Fixed this by setting `screen
= False` on the error callback of `getUserInfo`