Split user and aggregate queries
As shown in https://github.com/e-mission/e-mission-phone/issues/153, the
dashboard is currently slow because we generate aggregate metrics serially with
user metrics, and retrieving aggregate data is slow.

This restructuring allows us to quickly get responses to user queries, which
are all we need for the main cards, while allowing the aggregate reponses,
which are below the fold, to show up a bit later and be populated in the
background.

Promises are awesome.