Turn off autolaunch for the goals walkthrough
It conflicts with the popups from the referral process, and if we have two
popups at the same time, we end up unable to click one either of them.
Also try to avoid allowing people to apply the same update twice, which will
confuse the heck out of ionic deploy.