Create a "Fix Map" button for the diary list as well
We added "Fix Map" buttons to all the screens to fix loading issues.
However, @immesys has reported that they don't work on the list view.
This is probably because on the list view we reload data and fix the tiles at
the same time. We may be able to use $timeout after the load to fix the tiles
if that is the problem. But let us add a button to this screen for now anyway
and if it works, experiment with the screen load.

https://github.com/e-mission/e-mission-phone/issues/75