Handle the local: present, native: missing case by writing to native
This is a fix for part (2) of https://github.com/e-mission/e-mission-phone/issues/181.
If localStorage stores consent but native does not, we set the value in the
native and let the user know. This is consistent with
https://github.com/e-mission/e-mission-phone/issues/181#issuecomment-258757700

We still need to test and fix part (1) - i.e. why did the popup not show up?