Fix the check for the preventDefault
The `dragging`, `doubleClickZoom`, `scrollWheelZoom` etc properties are not
defined on `this`, which is an L.Handler. Instead, they are defined in the
config for the handler's map. So we need to use the get the config and check
the flags against it instead.

Without this fix, in the detail screen, the click event is propagated to the
underlying map even when a layer is clicked. This causes the
popup/ionicActionSheet to get closed immediately after open.

With this fix, that doesn't happen any more.