Show the safe/suck sheet only after the timestamp has been set
In the case in which there are multiple bins which need to be set, we first
popup an action sheet asking people to choose the time that the incident
happened. The safe/suck action sheet is supposed to be shown after that is
one.

However, because there was no dependency between the two sheets, they were
both displayed in parallel. And because the time selection sheet was shown
first, it was hidden by the safe/suck action sheet.

Fixed this by refactoring the safe/suck sheet into a `showSheet` function
which is called directly in the single time bin case, and after the time is
chosen in the multiple time bin case.