Add the newly added marker to the trip geojson
If we don't do this, then after adding the incident, it won't show up in the
list view, or on returning to the detail view until the list is refreshed.
This is because we have added this marker to the detail map, but the map is
deleted when we return to the list view, and the incidents are shown from the
trip geojson.

So we need to add the geojson corresponding to the incident to the geojson as
well.

Note that this required some restructuring of the code because we just have
the section in the incident list, and in order to get the trip, we need to
look it up from the timeline.

Adding a reference to the trip to the section causes slowness and crashes,
probably because of circular references (section -> trip -> section).

Looking up from the timeline directly causes a circular reference (Timeline -> PostTripManualMarker -> Timeline).

So we needed to read the markers in the controller, outside the timeline.