Move the incident pointFormat + onFeature to the incident class
This consolidates all the incident stuff in one place, and makes it possible
to reuse the behavior in other places like the heatmap. Also, create a
duplicate getFormattedTime method in the incident class. We can't use the one
from the diary handler because of circular dependencies.

TODO: Consider moving the formatting to a separate util class that can be used
in both to avoid repetition.