Refactor the adding incident code to take a set of points instead of a section
Although the code worked fine in the emulator, while testing on a real device,
I found that the target of the touch event was frequently the parent svg
instead of the actual <g> for the section. Since the listeners are bound to the
section, incidents were not added.

I am able to get the click to work if zoomed to the maximum extent, and if I am
extremely careful about exactly how I click, but it is not very reproducible,
and I am concerned that people won't know how it works. For example, I had to
click 5 times at the max zoom before I could get it to work.

On the other hand, maybe this is good because we really want precise
identification at the intersection level to make targeted improvements. On the
third hand, we are going to use auto detection instead of the user initiated
detection in most cases anyway.

Refactoring to take in a set of points instead of a section. Then we can reuse
the same code for map-level clicks and make this much easier to use.