Add a click handler to the map that will add an incident
As long as the point is within a small threshold of the actual trip points,
add an incident even when the map is clicked and not the section.  This
addresses the issue with click targets being correctly assigned on actual
mobile devices (see f1936155470a2ea5f1d13e3caa7190d4e98d6b29 for details).
This uses the newly restructured code from
f1936155470a2ea5f1d13e3caa7190d4e98d6b29 to add a new
`startAddingIncidentToTrip` and add it as a listener for the map.

Note that this adds a dependency between the diary controller and the incident
marker code.

It turns out that the distance threshold had to be different on iOS and
android to work correctly, which is why `DISTANCE_THRESHOLD` is a function
instead of a variable.