Upgrade the datepicker to the most recent version
This allows us to display the datepicker on a button click instead of using a
directive, which makes it much easier to use in the profile. We can't use the
most recent version v1.2.1 because it doesn't have support for the titleLabel.
We sync to a particular commit, which does have that support, instead.