Fix minor but annoying error while accessing consent in local storage
I was reading/writing the consent using
`startprefs.DATA_COLLECTION_CONSENTED_PROTOCOL`, but the variable was not in
`startprefs`, it was just

```
var DATA_COLLECTION_CONSENTED_PROTOCOL = 'data_collection_consented_protocol';
```

This appears to be true since the very first commit.
https://github.com/e-mission/e-mission-phone/commit/66b84b57d52c50c8483c8b64d9acbf08df40c4e0
so I am not sure how this ever worked :)

Maybe since we were both reading and writing with "undefined" as the key, it
somehow worked?