Add support for a new UnifiedDataLoader service
This will read data from both the local buffer and the remote server, and
combine them into a single deduped list. The server already returns data
combined from the cache and the timeseries.
https://github.com/shankari/e-mission-server/commit/7449e0585be62c78d211a3832bd3156f193a78c1

We should be able to use this to read information from the cache as well, if we
switch away from the timeline object to a standard query instead.