Handle notification is trigger while in the foreground
If a local notification is generated when the app is in the foreground, on iOS,
the notification is not delivered in the traditional way. Instead, it is passed
in to the `didReceiveLocalNotification` function.

In the localNotification plugin, this winds through a different codepath and finally gets here,

```
    NSTimeInterval timeInterval = [notification timeIntervalSinceLastTrigger];

    NSString* event = (timeInterval <= 1 && deviceready) ? @"trigger" : @"click";
    NSLog(@"event = %@", event);
```

and generates a `trigger` event event instead of a `click` event.

In the PR, we handle `trigger` events correctly as well.

The naive implementation would just call the same function for `trigger` as for
`click`, but that would be sub-optimal, since the user can dismiss the
notification in the `click` case, but is forced to go to the incident screen in
the `trigger` even if she did not have a trip to report.

So we handle the trigger case separately by showing a prompt and only going to
the incident reporting screen if the user clicks "Report".