Handle only the POST_TRIP prompt notifications
Before this change, the post-trip notification callbacks were for all
notifications. In particular, if we registered for callbacks before the data
collection was initialized, we would get the "New data collection terms"
notification and it would be delivered to the app because it was in the
foreground.

And then we would show the incident screen, but with nothing on it.

As we add more notifications, let us make sure that we only handle the incident
notifications here...