Initial working version
Very hacky, but at least works. The issue turns out to be that the document has
not actually loaded when `loadstop` is called. So I wait until the element is
non-null and then set it. This works just fine.

The current code also includes several alerts intended to help debug the
process as it happens!

This will definitely need to be cleaned up.