Move the uuid insertion script to a separate .js file
As the script gets more complex, sticking it into an ever-longer
inline "code" script gets way too complicated. Instead, since
we are reading in the script file manually and passing it in as 'code', we can
just put a placeholder in the script and replace it with the real value before
passing it in as inline code.

Note that I also tried to use the callback from the original `executeScript`
as described in http://stackoverflow.com/posts/23387583/revisions, but it did
not appear to work.