Add a new "mood question" icon to be used in the notification bar
Thanks to Manonmani and Jayalakshmi Raffill, create a new "mood question" icon
that works with the alpha channel squishing that android does for the
notification bar. This allows us to use it as the small icon, and have
something other than a white or black square show up in the android
notification bar.

Also check in a script that generates icons by copying them from the native
code directory (used for creating new icons) and a hook that automatically
copies existing icons to the native platform directory for people who just
check out the repository.