Work around back button not working by adding a close button
This fixes #206. There are issues with using either the popup method (hard to
specify custom controller) or the rootScope method (need to pass in state
params for this state but not for any other states).

The display works, there is just no way to go from it to another state. So I
just add a close button that returns to the profile/control state. Simple,
easy, not too much work given that we probably want to merge this with the
diary anyway long-term.