Don't handle notification triggers on android
On android, the notifications show up in the notification bar even if the app
is in the foreground. So we do not need to handle the 'trigger' case. And in
fact, handling it is bad because then if the app has been running in the
background, we end up with multiple invocations, one of each completed trip.