Reconfigure with the new server interval settings
during reconfigure. Note that this requires connectivity, but the prior code
did as well, so that's OK. We have this kick-ass sync infrastructure - we
should really use it and populate from the server side.

It is also worth thinking about where best to set this Currently happens in
native code. Now that we are switching away from parse, we can store this from
javascript here.  or continue to store from native this is easier for people to
see, but means that calls from other modules to native, even through the
javascript interface are not complete.