Move the standard launcher icon from `mipmap` to `drawable`
The most recent android versions store their icon under
`mipmap..." (e.g. mipmap-hdpi, mipmap-xhdpi, mipmap-xxxhdpi)
https://android-developers.googleblog.com/2014/10/getting-your-apps-ready-for-nexus-6-and.html

But that means that if we want to have our app icon as the large icon, we need
to have it in assets - only `drawables` are loaded otherwise. And if it an
asset - unsure if sizing and other things will work well.

So we add another hook that will copy `icon.png` from the `mipmap-*` directories to the corresponding `drawable` directories.