Use the newly promisifed code to read unprocessed data
We still need to convert the unprocessed data into trips but this seems to work
for reading unprocessed data, at least on a real android device.

It does not work in the iOS simulator - it sometimes works partially, but then
stops working with

```
[Log] DEBUG:while reading data for 1488787200000 from server, got nTrips = 0 (console-via-logger.js, line 173)
[Log] About to hide 'Reading from server' (console-via-logger.js, line 173)
[Log] Finished hiding ionicLoading, returning list of size 0 (console-via-logger.js, line 173)
```

Basically, it looks like the timeline call to the server completes, but the
call to get the processed ts does not, and so `Promise.all` is never invoked. I
plan to experiment further to get this to work reliably, but let's check in
this working version for now.