Mark trips as unprocessed (draft)
Use the functions added in 6880a1b60c6a73c24aaf5a192826a2e0eaf11807
to change the trip background to a dull green if it is a draft. Do this for
both the list and the detail view.

Also add `(draft)` text to indicate why those trips are so special, and add a
button to re-run the pipeline that currently shows a "coming soon" message.