Simply go to the diary when the incident is clicked
Instead of doing various contortions to go to the profile screen and pop things
up, we can simply go to the diary screen, now that unprocessed trips are
displayed.

This also allows us to go directly to the diary instead of metrics if launched
as part of an incident check.

Finally, if we go to an already loaded diary page because of an incident
report, we want to reload it to see the newest trip