Handle the case in which the server is down
If the server is down, we read data from the local cache.
But in this case, we weren't returning the promise properly.
Fixed it now, and tested by shutting down the dev server on my laptop

Also, switch the logger to the logger service, so we don't have a lot of
"undefined"s at the end of every message.