Support automatic redirection based on local notifications
Support automatically redirecting to a new state when the user clicks on a
local notification of a particular form. This is used to navigate to the
profile screen when the user needs to login
(see https://github.com/e-mission/cordova-jwt-auth/pull/13 and in particular,
https://github.com/e-mission/cordova-jwt-auth/pull/13/commits/cfa54573fb611d1c82feeeecf1ec6ca8720450ad)

In addition, reload the page when navigating back to the control screen so that
the user looks "logged in".