Fix toggle buttons
@krisma had used a special css for the `toggle-color` that always set that
color, whether the toggle was turned on or not. This meant that the toggle
background was always coloured, whether the toggle was on or not.

This meant that the toggle position was the only arbiter of what the setting
was, which was confusing. The standard ionic toggle was much better, because
it goes to a white background when the toggle is off.

This moves the toggle-color up to the label and ensures that the off/on case
is handled correctly.