Move the settings for the plugins to the plugin code
The files deleted from here are moved to the plugins in
https://github.com/e-mission/e-mission-data-collection/pull/153 and
https://github.com/e-mission/cordova-server-sync/pull/28

They will automatically be added back then the plugins are installed
(look at the `<asset>` tags in `plugin.xml`)

Using the settings are optional, so there is some glue code in here to ensure
that they are handled. This includes adding the appropriate script tags, and
adding some boilerplate code in control.js to update $scope variables when
the data is loaded.