Handle notification actions correctly in android
On iOS, actions other than report are handled in the background and don't
launch the app. But on android, there is no real support for background in
the plugin, so the app is launched. But then you have the app launch and not
do anything which means that people are unsure whether anything actually
happened.

Handle this by using $ionicPopups in the android case alone.