Fix the walkthrough for the list view
The list view walkthrough didn't work because if you have a nav-bar with an
overridden title, it shows up in both the cached and active views, and then
the element with the first id or class is not matched.

Fix this by moving the button to the nav-buttons, which are the ones that are
not duplicated. I guess title is supposed to be used only for text?

This fixes #229.

Also restore the navbar from the diary to be consistent with other views.