Fix zero length trips in unprocessed data
This is not a very common use case, but I encountered it during testing.
Basically, an unprocessed trip had zero points. While this should never
happen, we should also not crash if it does happen.

The fix is simple - we return an undefined trip if it has no points, and we
filter undefined trips before linking.