Move the connection settings (including keys and host information)
New location to a json file in the www directory.
This allows channels to host their own servers without deploying new apps.

Also move the startupConfig (which contains protocol information) to a sample
and update the README to indicate that the samples need to be copied to real
json files.