Fix the check for sync done
This worked when I first tested it (I am not sure why) but it didn't work when
I re-tried it with one of the CCI people present. The main challenge was that
we only push data until the last T_TRIP_ENDED transition, so any battery
readings after that would continue to accumulate in the usercache.

But it didn't work even when I manually stopped and started a trip. This is
because the sync at the end of a trip generates several battery messages,
all within a short time of each other. And these are AFTER the transition, so
they are not pushed at the end of the trip. And because we have 7 battery
readings, we assume that there is still data to push.

I fix this by moving to the transitions directly as a check. There are not many
transitions since they are only generated at the start and end of a trip (so
even less than battery readings) and because they are consistent with check on
when to sync, we know that when the last sync is done, we will return true.

Testing done:

Turned off the local server
Started and ended several trips
Forced sync -> pending sync
Started server
Forced sync -> sync done